[feat-12]: Create structs for saving the slots
