+++ pkg/config/deactivate.go
@@ -1,0 +1,6 @@
+package configs
+
+// Closing all connections, event listeners, etc
+func Deactivate() {
+	CloseDBConnection()
+}
