+++ pkg/consts/chain.const.go
@@ -1,0 +1,6 @@
+package consts
+
+
+const (
+	EpochLimit uint = 5
+)
