--- pkg/structs/common.struct.go
@@ -1,8 +1,0 @@
-package structs
-
-import "github.com/SidVermaS/Ethereum-Node-Indexer/pkg/models"
-
-type ValidatorToValidatorsStatusChannelStruct struct {
-	Validators        []*models.Validator
-	ValidatorStatuses []*models.ValidatorStatus
-}
